+++ lib/features/menu/presentation/contacts_screen.dart
@@ -1,5 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:collection/collection.dart';
+import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
 import 'package:zachranobed/common/constants.dart';
@@ -9,6 +10,7 @@ import 'package:zachranobed/extensions/build_context_extensions.dart';
 import 'package:zachranobed/features/menu/domain/model/contact.dart';
 import 'package:zachranobed/features/menu/domain/model/contacts_summary.dart';
 import 'package:zachranobed/features/menu/domain/usecase/get_contacts_use_case.dart';
+import 'package:zachranobed/ui/widgets/adaptive_content.dart';
 import 'package:zachranobed/ui/widgets/contact_row.dart';
 import 'package:zachranobed/ui/widgets/error_content.dart';
 import 'package:zachranobed/ui/widgets/menu/menu_section.dart';
@@ -34,6 +36,25 @@ class _ContactsScreenState extends State<ContactsScreen> {
 
   @override
   Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: AdaptiveContent(
+          web: (context) {
+            return Center(
+              child: SizedBox(
+                width: LayoutStyle.webBreakpoint.toDouble(),
+                child: _contactsScreenContent(),
+              ),
+            );
+          },
+          mobile: (context) => _contactsScreenContent(),
+        ),
+      ),
+    );
+  }
+
+  /// Builds the content of the contacts screen.
+  Widget _contactsScreenContent() {
     return Scaffold(
       appBar: AppBar(),
       body: FutureBuilder(